Added Grenade Layer
You could use this Layer on the grenades, so the playerr wont be shoved back when spawning them. But you cant kick them anymore. Have to work on better solution: GrenadeSpawnpoint outside of playerCollidor or maybe a scritp that activates the grenades collidor after a short time, so the collidor becomes only active once the grenade is otuside the palyers collidor.